+++ apps/recruitments/serializers/application_serializers.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
-from typing import Any, Dict, Optional
-
-from rest_framework import serializers
+from typing import Any, Optional
 
 from apps.recruitments.models.application import Application
+from rest_framework import serializers
 
+from apps.users.models import User
 
 class ApplicationCreateSerializer(serializers.ModelSerializer[Any]):
     """
@@ -49,3 +49,58 @@ class ApplicationResponseSerializer(serializers.Serializer[Any]):
     def get_user_uuid(self, obj: Application) -> Optional[str]:
         u = getattr(obj.user, "uuid", None)
         return str(u) if u else None
+
+class ApplicantSerializer(serializers.ModelSerializer[User]):
+    """
+    지원자 기본 정보 serializer
+    """
+
+    class Meta:
+        model = User
+        fields = ["nickname", "gender", "profile_img_url"]
+
+
+class ApplicationListSerializer(serializers.ModelSerializer[Application]):
+    """
+    지원 내역 목록 조회용 serializer
+    """
+
+    applicant = ApplicantSerializer(source="user", read_only=True)
+    applied_at = serializers.DateTimeField(source="created_at", format="%Y-%m-%d %H:%M", read_only=True)
+
+    class Meta:
+        model = Application
+        fields = [
+            "id",
+            "uuid",
+            "applicant",
+            "available_time",
+            "has_study_experience",
+            "status",
+            "applied_at",
+        ]
+
+
+class ApplicationDetailSerializer(serializers.ModelSerializer[Application]):
+    """
+    지원 내역 상세 조회용 serializer
+    """
+
+    applicant = ApplicantSerializer(source="user", read_only=True)
+    applied_at = serializers.DateTimeField(source="created_at", format="%Y-%m-%d %H:%M", read_only=True)
+
+    class Meta:
+        model = Application
+        fields = [
+            "id",
+            "uuid",
+            "applicant",
+            "self_introduction",
+            "motivation",
+            "objective",
+            "available_time",
+            "has_study_experience",
+            "study_experience",
+            "status",
+            "applied_at",
+        ]
